requirements.txt
asgiref
async-timeout
blinker
channels
channels-redis
click
Django
django-cors-headers
django-leaflet
django-redis
djangorestframework
djangorestframework-gis
djangorestframework-simplejwt
Flask
geos
importlib-metadata
itsdangerous
Jinja2
lxml
MarkupSafe
msgpack
Pillow
psycopg2-binary
PyJWT
python-decouple
pytz
redis
sqlparse
typing_extensions
Werkzeug
zipp
pyngrok==5.0.5.